+++ test/KMonad/ComposeSeqSpec.hs
@@ -18,5 +18,5 @@ spec = describe "compose-sequences" $ traverse_ checkComposeSeq ssComposed
   parsesAsValidComposeSeq (Right (KComposeSeq seq')) = all isSimple seq'
   parsesAsValidComposeSeq _ = False
   isSimple (KEmit _) = True
-  isSimple (KAround (KEmit KeyLeftShift) (KEmit _)) = True
+  isSimple (KAroundImplicit (KEmit KeyLeftShift) (KEmit _)) = True
   isSimple _ = False